DID alot!
The first part of the app, the salary section is almost done. A few things need to be done.

- I want the chart to have points that display the year that you get a promotion. So I think I have to create a dataframe with that year and salary increase, instead of a df with year and promotion number like I previously thought.

I work on this some other time, lol I gotta go to bed, lots of classes today... it's like 01:10

* I've been thinking that in the salary module, I can work create a simple taxation system, ofcourse that allows user inputs. That means in the input data part of the UI, I'll have to add a drop down menu where users can optionally enter more information.

It's looking good! More to be done, lots to learn.

Good Night,
Lloydinho signing out 🫡